Fix: refactor package of auth
